requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
cycler==0.11.0
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==2.2.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.2
nvidia-cublas==11.2.1.74
nvidia-cuda-cupti==11.1.69
nvidia-cuda-nvcc==11.1.74
nvidia-cuda-nvrtc==11.1.74
nvidia-cuda-runtime==11.1.74
nvidia-cudnn==8.0.4.30
nvidia-cufft==10.3.0.74
nvidia-curand==10.2.2.74
nvidia-cusolver==11.0.0.74
nvidia-cusparse==11.2.0.275
nvidia-dali-cuda110==0.26.0
nvidia-dali-nvtf-plugin==0.26.0+nv20.10
nvidia-nccl==2.7.8
nvidia-pyindex==1.0.5
nvidia-tensorboard==1.15.0+nv20.10
nvidia-tensorflow==1.15.4+nv20.10
nvidia-tensorrt==7.2.1.4
opt-einsum==3.3.0
pip==21.3.1
protobuf==3.19.6
psutil==6.1.1
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
setuptools==59.6.0
six==1.16.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tornado==6.1
typing_extensions==4.1.1
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.14.1
zipp==3.6.0
